backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
psycopg2-binary>=2.9,<3.0
langchain-core>=0.2,<1.0
email-validator>=2.1,<3.0
